seperate concerns in codebase create routes folder with auth event and categories routes create middleware folder with event and category validations create controller folders with controllers for auth events and categories
