+++ api/server.js
@@ -4,9 +4,7 @@ const cors = require('cors');
 const dotenv = require('dotenv');
 
 dotenv.config();
-const authRouter = require('../api/auth/authRouter');
-const eventsRouter = require('../api/events/eventsRouter');
-const eventCategoryRouter = require('../api/eventsCategories/eventCategoryRouter');
+const routes = require('../routes');
 
 const server = express();
 
@@ -14,9 +12,7 @@ server.use(helmet());
 server.use(express.json());
 server.use(cors());
 
-server.use('/api/auth', authRouter);
-server.use('/api/events', eventsRouter);
-server.use('/api/event-category', eventCategoryRouter);
+server.use('/api', routes);
 
 server.get('/', (req, res) => {
   res.status(200).json({
