+++ controllers/eventsCategories/eventsCategory.spec.js
@@ -1,5 +1,5 @@
 const request = require('supertest');
-const server = require('../server');
+const server = require('../../api/server');
 const db = require('../../data/dbConfig');
 
 let token;
