+++ routes/auth.js
@@ -1,14 +1,25 @@
-const express = require('express');
+const { Router } = require('express');
 const bcrypt = require('bcrypt');
 const passport = require('passport');
 const GoogleStrategy = require('passport-google-oauth').OAuth2Strategy;
 const GitHubStrategy = require('passport-github2').Strategy;
-const db = require('./authModel');
-const generateToken = require('../../utils/generateToken');
-const bodyValidator = require('../../utils/validator');
+const db = require('../controllers/auth/authModel');
+const bodyValidator = require('../utils/validator');
+const {
+  register,
+  Login,
+  getAuthToken
+} = require('../controllers/auth/authControllers');
 
-const router = express.Router();
-const server = require('../server');
+const router = Router();
+const server = require('../api/server');
+
+/**
+ * User Registration and Login Routes
+ */
+router.post('/register', bodyValidator, register);
+
+router.post('/login', bodyValidator, Login);
 
 // Passportjs config
 router.use(passport.initialize());
@@ -19,53 +30,6 @@ passport.serializeUser((user, done) => {
 
 passport.deserializeUser((user, done) => {
   done(null, user);
-});
-
-router.post('/register', bodyValidator, (req, res) => {
-  // endpoint to register
-  const newUser = req.body;
-  const hash = bcrypt.hashSync(newUser.password, 15);
-  newUser.password = hash;
-
-  db.addUser(newUser)
-    .then(user => {
-      const token = generateToken(user);
-      res.status(201).json({
-        user,
-        token
-      });
-    })
-    .catch(error => {
-      res.status(500).json({
-        message: `Couldnt register user: ${error.message}`,
-        data: error
-      });
-    });
-});
-
-router.post('/login', bodyValidator, (req, res) => {
-  // login endpoint
-  const { email, password } = req.body;
-
-  db.getUserBy({ email })
-    .then(user => {
-      if (user && bcrypt.compareSync(password, user.password)) {
-        const token = generateToken(user);
-        res.status(200).json({
-          token,
-          userId: user.id
-        });
-      } else {
-        res.status(400).json({
-          message: 'Invalid password!'
-        });
-      }
-    })
-    .catch(error => {
-      res.status(500).json({
-        message: `Internal server error${error.message}`
-      });
-    });
 });
 
 // google
@@ -171,33 +135,6 @@ router.get(
   }
 );
 
-router.get('/token', async (req, res) => {
-  try {
-    const data = server.locals;
-    if (!data) {
-      res.status(400).json({
-        statusCode: 400,
-        message: 'Authentication Failed'
-      });
-    }
-    const user = await db.createOrFindUser(data);
-    if (user) {
-      req.user = server.locals;
-      const token = generateToken(user);
-      res.status(200).json({
-        statusCode: 200,
-        message: `${req.user.authType} Login was successfull`,
-        token,
-        userDetails: {
-          username: user.username,
-          email: user.email,
-          fullname: user.fullname
-        }
-      });
-    }
-  } catch (error) {
-    return error;
-  }
-});
+router.get('/token', getAuthToken);
 
 module.exports = router;
